restructure components tests
