+++ tests/test_components.py
@@ -36,21 +36,3 @@ class TestComponents(unittest.TestCase):
     @pytest.fixture(autouse=True)
     def _use_caplog(self, caplog):
         self.caplog = caplog
-
-    ############################################
-    # Test of the individual components commands. #
-    ############################################
-
-    from .components.generate_snapshot import (  # type: ignore[misc]
-        test_generate_snapshot_module,
-        test_generate_snapshot_once,
-        test_generate_snapshot_subworkflow,
-        test_test_not_found,
-        test_unstable_snapshot,
-        test_update_snapshot_module,
-    )
-    from .components.snapshot_test import (  # type: ignore[misc]
-        test_components_test_check_inputs,
-        test_components_test_no_installed_modules,
-        test_components_test_no_name_no_prompts,
-    )
